revert: fix: add scroll on overflow for walletconnect to dapps menu (#5406)
This reverts commit 4fa4e57cc4492c77690ecbbb77e148e0ea60ec0d.
